refactor database to expose db setup for testing
